requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow>=10.0.0
python-decouple==3.8
django-filter==23.5

# Database drivers (optional - only needed if using MySQL or PostgreSQL)
# For MySQL, choose one:
# mysqlclient>=2.2.0  # Recommended, but may require compilation
# mysql-connector-python>=8.0.0  # Pure Python alternative (used by setup_mysql.py)

# For PostgreSQL:
# psycopg2-binary>=2.9.0  # Recommended for PostgreSQL

# SQLite is included with Python, no additional driver needed
